split package into modules
