--- meer_tec/__init__.py
@@ -1,1 +1,0 @@
-from .meer_tec import TEC, USB, Message, XPort  # noqa: F401
